+++ src/main/java/chess/api/storage/ephemeral/IntArrayComparator.java
@@ -2,10 +2,10 @@ package chess.api.storage.ephemeral;
 
 import java.util.Comparator;
 
-public class IntArrayComparator implements Comparator<int[]> {
+public class IntArrayComparator implements Comparator<short[]> {
 
     @Override
-    public int compare(int[] t1, int[] t2) {
+    public int compare(short[] t1, short[] t2) {
         if (t1.length == 0 && t2.length == 0) {
             return 0;
         } else if (t1.length == 0) {
@@ -17,8 +17,8 @@ public class IntArrayComparator implements Comparator<int[]> {
             if (index >= t2.length) {
                 return 1;
             }
-            int val1 = t1[index];
-            int val2 = t2[index];
+            short val1 = t1[index];
+            short val2 = t2[index];
             if (val1 < val2) {
                 return -2;
             } else if (val1 > val2) {
